refactor: move metrics helper in lib
