+++ lib/cpu.js
@@ -5,7 +5,7 @@ const {
   kComputedMetric,
   kSample,
   kReset
-} = require('../symbols')
+} = require('./symbols')
 
 const kCpuLastSample = Symbol('doc.metrics.cpu.lastSample')
