+++ lib/eventLoopDelay.js
@@ -8,7 +8,7 @@ const {
   kOptions,
   kReset,
   kSample
-} = require('../symbols')
+} = require('./symbols')
 
 class EventLoopDelayMetric {
   constructor (opts) {
