+++ lib/gc.js
@@ -20,7 +20,7 @@ const {
   kReset,
   kComputedMetric,
   kObserverCallback
-} = require('../symbols')
+} = require('./symbols')
 
 const flagsSupported = NODE_PERFORMANCE_GC_FLAGS_NO !== undefined
